rework init sequence
Send a soft reset (AT&F) and wait for any startup chatter to finish.
If an OK response is not received, send a reset command.  If this is
not accepted, the hardware watchdog will eventually reset the modem.

These changes avoid relying on the buffered backlog of messages being
sent on connect.  Apparently this is not reliable.