include save flag in gpio commands only when supported
On the SIM5360, gpio commands include a flag indicating whether to
save the state of the pin, defaulting to 1.  We want to clear this
flag so as to avoid wearing out nonvolatile memory when toggling
the watchdog pin.

The SIM7600 doesn't support this flag at all.  Thus check the model
and append the flag when it is known to be supported.